feat: move files and change namespaces
