+++ aplib.net-demo/Assets/Scripts/Items/SpawnWave.cs
@@ -1,4 +1,3 @@
-using Assets.Scripts.Wfc;
 using LevelGeneration;
 using UnityEngine;
