+++ aplib.net-demo/Assets/Scripts/LevelGeneration/SpawningExtensions.cs.meta
